test: drop the request-timeout half, keep the backup test
Reverts internal/request/request.go and the notification test to main
exactly, and removes the timeout tests.

The remaining assertions did not earn the production change they needed.
TestCallClientTimeoutIsFixed compared a constant against a literal in
the same file, and TestCallReturnsWhenTimeoutFires proved that Call uses
the package-level client rather than building its own -- a single line of
indirection. Neither justifies moving production code for a test's
benefit, which was the objection in the first place.

TestSlackNotification_HangingReceiverShouldTimeOut goes back to being
skipped. Asserting the real 30s bound costs a 30s wall-clock wait, and
that trade was the right call when it was made.

What is left is the half that was worth the PR: the backup happy path,
gated behind BLNK_TEST_PG_DUMP -- which nothing in the repo sets -- now
runs against a pg_dump stub and asserts a non-empty dump lands under
BackupDir.
